backend/server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cached-property==2.0.1
click==8.3.1
clikit==0.6.2
cmdstanpy==1.3.0
comm==0.2.3
contourpy==1.3.3
crashtest==0.3.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
frozenlist==1.8.0
holidays==0.85
httpstan==4.13.0
idna==3.11
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pastel==0.2.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.2.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pylev==1.4.0
pyparsing==3.2.5
pysimdjson==6.0.2
pystan==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
sqlalchemy-orm==1.2.10
stack-data==0.6.3
stanio==0.5.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
webargs==8.7.1
Werkzeug==3.1.3
yarl==1.22.0
